Reverted changes to Registry. - Must've been the wind...
