+++ loaforcsSoundAPI/Core/JSON/JSONDataLoader.cs
@@ -115,24 +115,6 @@ public static class JSONDataLoader {
 
 			return property;
 		}
-
-		protected override JsonConverter ResolveContractConverter(Type objectType) {
-			TypeInfo typeInfo = objectType.GetTypeInfo();
-
-			JsonConverterAttribute jsonConverterAttribute = typeInfo.GetCustomAttribute<JsonConverterAttribute>();
-			TypeInfo converterTypeInfo = jsonConverterAttribute?.ConverterType?.GetTypeInfo();
-
-			if(converterTypeInfo?.IsGenericTypeDefinition == true) {
-				while(typeInfo != null) {
-					if(typeInfo.IsGenericType && typeInfo.GetGenericTypeDefinition() == typeof(Registry<,>)) {
-						return (JsonConverter) Activator.CreateInstance(converterTypeInfo.MakeGenericType([.. typeInfo.GenericTypeArguments, objectType]), jsonConverterAttribute.ConverterParameters);
-					}
-					typeInfo = typeInfo.BaseType?.GetTypeInfo();
-				}
-			}
-
-			return base.ResolveContractConverter(objectType);
-		}
 	}
 
 	class ConditionConverter : JsonConverter<Condition> {
