+++ loaforcsSoundAPI/SoundPacks/Data/Registry.cs
@@ -1,15 +1,20 @@
-﻿using System;
-using System.Collections;
+﻿using System.Collections;
 using System.Collections.Generic;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
+using System.IO;
+using System.Linq;
+using loaforcsSoundAPI.Core;
+using loaforcsSoundAPI.Core.Data;
+using loaforcsSoundAPI.Core.JSON;
+using loaforcsSoundAPI.Core.Util;
 
 namespace loaforcsSoundAPI.SoundPacks.Data;
 
-[JsonConverter(typeof(RegistryConverter<,,>))]
-public abstract class Registry<T, TCollection>() : IEnumerable<T> where TCollection : ICollection<T>, new() {
-	readonly TCollection _items = []; // Generic ICollection holding entries in this Registry.
-	internal JToken[] _temp; // Array holding JSON tokens to be parsed, in case it needs to be delayed.
+public class Registry<T>(SoundPack pack, string relativePath) : IEnumerable<T> where T : class, IFilePathAware {
+	readonly List<T> _items = [];
+
+	public SoundPack Pack { get; } = pack;
+	public string RelativePath { get; } = relativePath;
+	public string AbsolutePath { get; } = Path.Combine(pack.PackFolder, relativePath);
 
 	public IEnumerator<T> GetEnumerator() {
 		return _items.GetEnumerator();
@@ -19,53 +24,85 @@ public abstract class Registry<T, TCollection>() : IEnumerable<T> where TCollect
 		return GetEnumerator();
 	}
 
-	public virtual void PopulateRegistry() {
-		for(int i = 0; i < _temp?.Length; i++) {
-			if(TryParse(out T value, _temp[i])) {
-				AddValue(value);
+    T TryLoadFile(string filePath) {
+        T item = JSONDataLoader.LoadFromFile<T>(filePath);
+        if(item == null) return default; // json error
+
+        if(item is IPackData pd) {
+            pd.Pack = Pack;
+        }
+
+        if(item is IValidatable validatable) {
+            if(!IValidatable.LogAndCheckValidationResult(
+                   $"loading '{LogFormats.FormatFilePath(filePath)}",
+                   validatable.Validate(),
+                   Pack.Logger
+               )) {
+                return default;
+            }
+        }
+
+        if(item is IRegistrationCallback callback) {
+            callback.OnRegistered();
+        }
+
+        _items.Add(item);
+        return item;
+    }
+
+	internal void Load() {
+		if(!Directory.Exists(AbsolutePath)) return; // nothing to load!
+
+		foreach(string file in Directory.GetFiles(AbsolutePath, "*.json", SearchOption.AllDirectories)) {
+			TryLoadFile(file);
+		}
+	}
+
+	protected virtual void HotLoadAdd(T item) { }
+	protected virtual void HotLoadRemove(T item) { }
+
+	internal void EnableHotReload() {
+		FileSystemWatcher watcher = new FileSystemWatcher(AbsolutePath) {
+			NotifyFilter = NotifyFilters.Attributes | NotifyFilters.CreationTime | NotifyFilters.FileName | NotifyFilters.LastWrite,
+			Filter = "*.json",
+			IncludeSubdirectories = true
+		};
+
+		void UnloadOldFile(string fullPath) {
+			T existing = _items.FirstOrDefault(it => it.FilePath == fullPath);
+			if(existing != null) {
+				Debuggers.HotReload?.Log("\twas in registry, removing!");
+				_items.Remove(existing);
+				HotLoadRemove(existing);
 			}
 		}
-		_temp = null;
-		OnRegistryPopulated();
-	}
 
-	public virtual void OnRegistryPopulated() { }
-
-	public virtual bool TryParse(out T value, JToken token) {
-		value = token.ToObject<T>();
-		return value != null;
-	}
-
-	public virtual void AddValue(T value) => _items.Add(value);
-	public virtual bool RemoveValue(T value) => _items.Remove(value);
-	public virtual bool ContainsValue(T value) => _items.Contains(value);
-	public virtual T FindValue(Predicate<T> match) {
-		foreach(T value in _items) {
-			if(match(value)) {
-				return value;
+		void LoadNewFile(string fullPath) {
+			T item = TryLoadFile(fullPath);
+			if(item != null) {
+				HotLoadAdd(item);
 			}
 		}
-		return default;
-	}
-}
 
-class RegistryConverter<T, TCollection, TRegistry> : JsonConverter<TRegistry> where TCollection : ICollection<T>, new() where TRegistry : Registry<T, TCollection>, new() {
-	public override TRegistry ReadJson(JsonReader reader, Type objectType, TRegistry existingValue, bool hasExistingValue, JsonSerializer serializer) {
-		JToken token = JToken.Load(reader);
+		watcher.Created += (s, e) => {
+			Debuggers.HotReload?.Log($"Created: {LogFormats.FormatFilePath(e.FullPath)}");
+			LoadNewFile(e.FullPath);
+		};
+		watcher.Deleted += (s, e) => {
+			Debuggers.HotReload?.Log($"Deleted: {LogFormats.FormatFilePath(e.FullPath)}");
+			UnloadOldFile(e.FullPath);
+		};
+		watcher.Changed += (s, e) => {
+			Debuggers.HotReload?.Log($"Changed: {LogFormats.FormatFilePath(e.FullPath)}");
+			UnloadOldFile(e.FullPath);
+			LoadNewFile(e.FullPath);
+		};
+		watcher.Renamed += (s, e) => {
+			Debuggers.HotReload?.Log($"Renamed: {LogFormats.FormatFilePath(e.OldName)} -> {LogFormats.FormatFilePath(e.Name)}");
+			UnloadOldFile(e.OldFullPath);
+			LoadNewFile(e.FullPath);
+		};
 
-		if(token.Type is JTokenType.String) {
-			return new TRegistry() { _temp = [token] };
-		}
-
-		if(token is JArray jsonArray) {
-			return new TRegistry() { _temp = [.. jsonArray] };
-		}
-
-		IJsonLineInfo lineInfo = reader as IJsonLineInfo;
-		throw new JsonReaderException($"Value '{token}' is not valid for a registry.", reader.Path, lineInfo?.LineNumber ?? 0, lineInfo?.LinePosition ?? 0, null);
-	}
-
-	public override void WriteJson(JsonWriter writer, TRegistry value, JsonSerializer serializer) {
-		throw new NotImplementedException("no.");
+		watcher.EnableRaisingEvents = true;
 	}
 }
